Django.Meetup.Bulgaria-21-06-2022/demo/django/app/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.5
django-environ==0.8.1
psycopg2==2.9.2
sqlparse==0.4.2
djangorestframework==3.13.0
